requirements.txt
flask~=2.3.2
flask_sqlalchemy
psycopg2-binary
PyJWT~=2.7.0
SQLAlchemy==2.0.16
Werkzeug==2.3.6

